スタブの不具合を修正。

+++ Assets/UnityTetris/Tests/Stubs/StubField.cs
@@ -14,7 +14,7 @@ public class StubField : AbstractField
     public bool ReturnSetBlocks;
     public int ReturnWidth;
     public int ReturnHeight;
-    public float[,] ReturnGetFieldMap; 
+    public int[,] ReturnGetFieldMap; 
 
     public void ClearCallList()
     {
@@ -63,7 +63,7 @@ public class StubField : AbstractField
         CallList += $"ReduceLines({owner.GetHashCode()})\n";
     }
 
-    public override float[,] GetFieldMap()
+    public override int[,] GetFieldMap()
     {
         CallList += "GetFieldMap\n";
         return ReturnGetFieldMap;
